backend/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "hivemind"
version = "2.0.0"
description = "Production-grade multi-agent AI orchestration system"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
authors = [
    { name = "HIVEMIND Contributors" }
]
keywords = ["ai", "agents", "orchestration", "claude", "openai", "multi-agent"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Framework :: FastAPI",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Typing :: Typed",
]

dependencies = [
    # Core Framework
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",

    # Async Support
    "anyio>=4.2.0",
    "httpx>=0.26.0",

    # Database - Relational
    "sqlalchemy[asyncio]>=2.0.25",
    "asyncpg>=0.29.0",
    "alembic>=1.13.0",

    # Database - Vector
    "qdrant-client>=1.7.0",

    # IPC - Message Queue
    "pyzmq>=25.1.0",
    "redis>=5.0.0",
    "aio-pika>=9.3.0",

    # Serialization
    "msgpack>=1.0.7",
    "orjson>=3.9.0",

    # Resilience
    "tenacity>=8.2.0",
    "pybreaker>=1.0.0",

    # Security
    "hvac>=2.1.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",

    # Observability
    "structlog>=24.1.0",
    "prometheus-client>=0.19.0",
    "opentelemetry-api>=1.22.0",
    "opentelemetry-sdk>=1.22.0",
    "opentelemetry-exporter-otlp>=1.22.0",
    "opentelemetry-instrumentation-fastapi>=0.43b0",
    "opentelemetry-instrumentation-sqlalchemy>=0.43b0",
    "opentelemetry-instrumentation-redis>=0.43b0",

    # gRPC
    "grpcio>=1.60.0",
    "grpcio-tools>=1.60.0",

    # CLI
    "typer>=0.9.0",
    "rich>=13.7.0",

    # Utilities
    "python-dotenv>=1.0.0",
    "toml>=0.10.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.12.0",
    "httpx>=0.26.0",
    "ruff>=0.1.9",
    "mypy>=1.8.0",
    "pre-commit>=3.6.0",
    "types-redis>=4.6.0",
    "types-toml>=0.10.8",
]

test = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.12.0",
    "httpx>=0.26.0",
    "testcontainers>=3.7.0",
    "factory-boy>=3.3.0",
    "faker>=22.0.0",
]

[project.scripts]
hivemind = "hivemind.cli.main:app"
hivemind-server = "hivemind.api.server:main"

[project.urls]
Homepage = "https://github.com/Nerds489/HIVEMIND"
Documentation = "https://github.com/Nerds489/HIVEMIND#readme"
Repository = "https://github.com/Nerds489/HIVEMIND"
Issues = "https://github.com/Nerds489/HIVEMIND/issues"

[tool.hatch.build.targets.wheel]
packages = ["src/hivemind"]

[tool.hatch.build.targets.sdist]
include = [
    "/src",
    "/tests",
    "/proto",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = [
    "-ra",
    "-q",
    "--strict-markers",
    "--cov=src/hivemind",
    "--cov-report=term-missing",
]
markers = [
    "slow: marks tests as slow",
    "integration: marks tests as integration tests",
]

[tool.ruff]
target-version = "py311"
line-length = 100
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "ARG",    # flake8-unused-arguments
    "SIM",    # flake8-simplify
]
ignore = [
    "E501",   # line too long (handled by formatter)
    "B008",   # do not perform function calls in argument defaults
    "C901",   # too complex
]

[tool.ruff.per-file-ignores]
"tests/*" = ["ARG"]

[tool.ruff.isort]
known-first-party = ["hivemind"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = true
plugins = ["pydantic.mypy"]

[[tool.mypy.overrides]]
module = [
    "pyzmq.*",
    "aio_pika.*",
    "hvac.*",
    "pybreaker.*",
    "msgpack.*",
    "grpc.*",
    "qdrant_client.*",
]
ignore_missing_imports = true

[tool.coverage.run]
source = ["src/hivemind"]
branch = true

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if TYPE_CHECKING:",
    "if __name__ == .__main__.:",
]
